+++ src/release.test.js
@@ -1,14 +1,15 @@
 import { defaultConfig } from './'
 import { changelog } from './changelog'
 import { release } from './release'
-import { commitRelease, updateVersion } from './utils'
+import { commitRelease, defineVersion, updateVersion } from './utils'
 
 jest.mock('./changelog', () => ({
   changelog: jest.fn()
 }))
 jest.mock('./utils', () => ({
   commitRelease: jest.fn(),
-  updateVersion: jest.fn()
+  updateVersion: jest.fn(),
+  defineVersion: jest.fn()
 }))
 jest.mock('console', () => ({
   error: jest.fn()
@@ -21,19 +22,17 @@ describe('release', () => {
     expect(release()).rejects.toThrow('Relese requires a version')
   })
 
-  it('should throw if incorrect version passed', () => {
-    const mockedVersion = 'version'
-
-    expect(release(mockedVersion)).rejects.toThrow(
-      "Version 'version' doesn't look right"
-    )
-  })
-
   it('should execute release', async () => {
     const mockedVersion = '2.2.2'
 
+    defineVersion.mockImplementation(() => mockedVersion)
+
     await release(mockedVersion, {}, defaultConfig)
 
+    expect(defineVersion).toBeCalledTimes(1)
+    expect(defineVersion).toBeCalledWith(mockedVersion)
+    expect(updateVersion).toBeCalledTimes(1)
+    expect(updateVersion).toBeCalledWith(mockedVersion)
     expect(changelog).toBeCalledTimes(1)
     expect(changelog).toBeCalledWith(
       mockedVersion,
@@ -42,7 +41,5 @@ describe('release', () => {
     )
     expect(commitRelease).toBeCalledTimes(1)
     expect(commitRelease).toBeCalledWith(mockedVersion)
-    expect(updateVersion).toBeCalledTimes(1)
-    expect(updateVersion).toBeCalledWith(mockedVersion)
   })
 })