+++ src/changelog.js
@@ -12,7 +12,12 @@ export const changelog = async (version, options, config) => {
   const latestTag = hasChangelog && !options.root && (await getLatestTag())
   const commits = await getCommits(latestTag)
 
-  if (!commits.length) throw new Error('No commits found since the last tag')
+  if (!commits.length) {
+    const message = latestTag
+      ? `No commits found since the latest tag '${latestTag}'`
+      : 'No commits found'
+    throw new Error(message)
+  }
 
   const grouped = groupCommits(commits, config)
   const changelog = generateChangelog(version, grouped, config)