+++ src/changelog.test.js
@@ -1,4 +1,4 @@
-import { existsSync, writeFileSync } from 'fs'
+import { writeFileSync } from 'fs'
 import { mockProcessStdout } from 'jest-mock-process'
 import { defaultChangelogOptions as defaultOptions, defaultConfig } from './'
 import { changelog } from './changelog'
@@ -19,8 +19,7 @@ jest.mock('./utils', () => ({
 }))
 jest.mock('fs', () => ({
   writeFileSync: jest.fn(),
-  readFile: jest.fn(),
-  existsSync: jest.fn()
+  readFile: jest.fn()
 }))
 
 describe('changelog', () => {
@@ -29,7 +28,6 @@ describe('changelog', () => {
   it('should throw if no commits found', async () => {
     const mockedCommits = []
 
-    existsSync.mockImplementation(() => true)
     getCommits.mockImplementation(() => mockedCommits)
 
     expect(changelog(null, defaultOptions, defaultConfig)).rejects.toThrow(
@@ -41,7 +39,6 @@ describe('changelog', () => {
     const mockedTag = '2.2.3'
     const mockedCommits = []
 
-    existsSync.mockImplementation(() => true)
     getLatestTag.mockImplementation(() => mockedTag)
     getCommits.mockImplementation(() => mockedCommits)
 
@@ -77,7 +74,6 @@ describe('changelog', () => {
       '# Latest\n\n## Features\n\n- add %HASH% placeholder to line format a3c93b2f\n- introduce changelog customization using config file e66d6176\n- use higher level of headers for changelog eea23d95\n\n## Fixes\n\n- replace %message% as last to avoid bugs ec507396\n- stop adding empty line at the end of the file on --root faee4801\n- stop adding Latest when not applicable c64fa467\n\n## Misc\n\n- include commit links in the changelog 8f622021\n\n'
     const mockedReleased = ''
 
-    existsSync.mockImplementation(() => true)
     getLatestTag.mockImplementation(() => mockedTag)
     getCommits.mockImplementation(() => mockedCommits)
     groupCommits.mockImplementation(() => mockedGrouped)
@@ -86,8 +82,6 @@ describe('changelog', () => {
 
     await changelog(null, defaultOptions, defaultConfig)
 
-    expect(existsSync).toBeCalledTimes(1)
-    expect(existsSync).toBeCalledWith('CHANGELOG.md')
     expect(getLatestTag).toBeCalledTimes(1)
     expect(getLatestTag).toBeCalledWith()
     expect(getCommits).toBeCalledTimes(1)
@@ -148,33 +142,6 @@ describe('changelog', () => {
     expect(writeFileSync).not.toBeCalled()
   })
 
-  it('should generate changelog with --root option', async () => {
-    const stdoutMock = mockProcessStdout()
-    const options = { ...defaultOptions, root: true }
-    const mockedTag = '2.2.3'
-    const mockedChangelog =
-      '# Latest\n\n## Features\n\n- add %HASH% placeholder to line format a3c93b2f\n- introduce changelog customization using config file e66d6176\n- use higher level of headers for changelog eea23d95\n\n## Fixes\n\n- replace %message% as last to avoid bugs ec507396\n- stop adding empty line at the end of the file on --root faee4801\n- stop adding Latest when not applicable c64fa467\n\n## Misc\n\n- include commit links in the changelog 8f622021\n\n'
-    const mockedCommits = [
-      'f2191200bf7b6e5eec3d61fcef9eb756e0129cfb chore(release): 0.1.0',
-      'aa805ce71ee103965ce3db46d4f6ed2658efd08d feat: add option to write to local CHANGELOG file',
-      '4e02179cae1234d7083036024080a3f25fcb52c2 feat: add execute release feature',
-      'bffc2f9e8da1c7ac133689bc9cd14494f3be08e3 refactor: extract line generating logic to function and promisify exec',
-      '2ea04355c1e81c5088eeabc6e242fb1ade978524 chore(changelog): update CHANGELOG'
-    ]
-    const mockedReleased = ''
-
-    getLatestTag.mockImplementation(() => mockedTag)
-    getCommits.mockImplementation(() => mockedCommits)
-    generateChangelog.mockImplementation(() => mockedChangelog)
-    generateReleased.mockImplementation(() => mockedReleased)
-
-    await changelog(null, options, defaultConfig)
-
-    expect(stdoutMock).toBeCalledTimes(1)
-    expect(stdoutMock).toBeCalledWith(mockedChangelog)
-    expect(writeFileSync).not.toBeCalled()
-  })
-
   it('should write to file with --write option', async () => {
     const options = { ...defaultOptions, write: true }
     const mockedTag = '2.2.3'
@@ -219,7 +186,6 @@ describe('changelog', () => {
     const mockedReleased = '# 2.2.2\n- feat: add feature 2da21c56'
     const mockedFilename = 'CHANGELOG.md'
 
-    existsSync.mockImplementation(() => true)
     getLatestTag.mockImplementation(() => mockedTag)
     getCommits.mockImplementation(() => mockedCommits)
     generateChangelog.mockImplementation(() => mockedChangelog)
@@ -235,5 +201,55 @@ describe('changelog', () => {
       mockedFilename,
       mockedChangelog + mockedReleased
     )
+  })
+
+  it('should generate changelog with --root option', async () => {
+    const stdoutMock = mockProcessStdout()
+    const options = { ...defaultOptions, root: true }
+    const mockedTag = '2.2.3'
+    const mockedChangelog =
+      '# Latest\n\n## Features\n\n- add %HASH% placeholder to line format a3c93b2f\n- introduce changelog customization using config file e66d6176\n- use higher level of headers for changelog eea23d95\n\n## Fixes\n\n- replace %message% as last to avoid bugs ec507396\n- stop adding empty line at the end of the file on --root faee4801\n- stop adding Latest when not applicable c64fa467\n\n## Misc\n\n- include commit links in the changelog 8f622021\n\n'
+    const mockedCommits = [
+      'f2191200bf7b6e5eec3d61fcef9eb756e0129cfb chore(release): 0.1.0',
+      'aa805ce71ee103965ce3db46d4f6ed2658efd08d feat: add option to write to local CHANGELOG file',
+      '4e02179cae1234d7083036024080a3f25fcb52c2 feat: add execute release feature',
+      'bffc2f9e8da1c7ac133689bc9cd14494f3be08e3 refactor: extract line generating logic to function and promisify exec',
+      '2ea04355c1e81c5088eeabc6e242fb1ade978524 chore(changelog): update CHANGELOG'
+    ]
+
+    getLatestTag.mockImplementation(() => mockedTag)
+    getCommits.mockImplementation(() => mockedCommits)
+    generateChangelog.mockImplementation(() => mockedChangelog)
+
+    await changelog(null, options, defaultConfig)
+
+    expect(stdoutMock).toBeCalledTimes(1)
+    expect(stdoutMock).toBeCalledWith(mockedChangelog)
+    expect(writeFileSync).not.toBeCalled()
+  })
+
+  it('should generate changelog with --from option', async () => {
+    const stdoutMock = mockProcessStdout()
+    const options = { ...defaultOptions, from: '4e02179c' }
+    const mockedTag = '2.2.3'
+    const mockedChangelog =
+      '# Latest\n\n## Features\n\n- add %HASH% placeholder to line format a3c93b2f\n- introduce changelog customization using config file e66d6176\n- use higher level of headers for changelog eea23d95\n\n## Fixes\n\n- replace %message% as last to avoid bugs ec507396\n- stop adding empty line at the end of the file on --root faee4801\n- stop adding Latest when not applicable c64fa467\n\n## Misc\n\n- include commit links in the changelog 8f622021\n\n'
+    const mockedCommits = [
+      'f2191200bf7b6e5eec3d61fcef9eb756e0129cfb chore(release): 0.1.0',
+      'aa805ce71ee103965ce3db46d4f6ed2658efd08d feat: add option to write to local CHANGELOG file',
+      '4e02179cae1234d7083036024080a3f25fcb52c2 feat: add execute release feature',
+      'bffc2f9e8da1c7ac133689bc9cd14494f3be08e3 refactor: extract line generating logic to function and promisify exec',
+      '2ea04355c1e81c5088eeabc6e242fb1ade978524 chore(changelog): update CHANGELOG'
+    ]
+
+    getLatestTag.mockImplementation(() => mockedTag)
+    getCommits.mockImplementation(() => mockedCommits)
+    generateChangelog.mockImplementation(() => mockedChangelog)
+
+    await changelog(null, options, defaultConfig)
+
+    expect(stdoutMock).toBeCalledTimes(1)
+    expect(stdoutMock).toBeCalledWith(mockedChangelog)
+    expect(writeFileSync).not.toBeCalled()
   })
 })